# docs/swagger.yaml
basePath: /
definitions:
...
  metadata.Actor:
    properties:
      name:
        type: string
      order:
        type: string
      role:
        type: string
      thumb:
        type: string
    type: object
  metadata.CastCrew:
    properties:
      actors:
        items:
          $ref: '#/definitions/metadata.Actor'
        type: array
      credits:
        items:
          type: string
        type: array
      directors:
        items:
          type: string
        type: array
    type: object
  metadata.DocumentKind:
    enum:
    - movie
    - tvshow
    - episodedetails
    - musicvideo
    - url
    - unknown
    type: string
    x-enum-varnames:
    - KindMovie
    - KindTVShow
    - KindEpisode
    - KindMusicVideo
    - KindURL
    - KindUnknown
  metadata.EpisodeBookmark:
    properties:
      position:
        type: string
    type: object
  metadata.EpisodeFields:
    properties:
      display_episode:
        type: string
      display_season:
        type: string
      episode:
        type: string
      episode_bookmark:
        $ref: '#/definitions/metadata.EpisodeBookmark'
      season:
        type: string
      show_title:
        type: string
    type: object
  metadata.Fanart:
    properties:
      thumbs:
        items:
          $ref: '#/definitions/metadata.Thumb'
        type: array
    type: object
  metadata.Link:
    properties:
      default:
        description: |-
          Default mirrors the default="true" attribute on a <uniqueid> tag, which
          names the provider a scraper treats as authoritative. Nothing in Metarr
          reads it yet; it is carried so a file we rewrite comes back as it went in.
        type: boolean
      key:
        type: string
      value:
        type: string
    type: object
  metadata.Metadata:
    properties:
      aired:
        type: string
      cast_crew:
        $ref: '#/definitions/metadata.CastCrew'
      code:
        type: string
      date_added:
        $ref: '#/definitions/types.Date'
      episode:
        $ref: '#/definitions/metadata.EpisodeFields'
      external_links:
        items:
          $ref: '#/definitions/metadata.Link'
        type: array
      extra:
        items:
          $ref: '#/definitions/metadata.UnknownElement'
        type: array
      genres:
        items:
          type: string
        type: array
      id:
        type: string
      kind:
        $ref: '#/definitions/metadata.DocumentKind'
      last_played:
        type: string
      movie:
        $ref: '#/definitions/metadata.MovieFields'
      mpaa:
        type: string
      musicvideo:
        $ref: '#/definitions/metadata.MusicVideoFields'
      original_title:
        type: string
      outline:
        type: string
      parse_error:
        description: |-
          ParseError records why an NFO could not be read. A malformed sidecar is
          reported here rather than failing the scan.
        type: string
      play_count:
        type: integer
      plot:
        type: string
      premiered:
        $ref: '#/definitions/types.Date'
      ratings:
        $ref: '#/definitions/metadata.Ratings'
      resume:
        $ref: '#/definitions/metadata.Resume'
      runtime:
        type: string
      scope:
        type: string
      status:
        type: string
      studios:
        items:
          type: string
        type: array
      tagline:
        type: string
      tags:
        items:
          type: string
        type: array
      target_base_name:
        type: string
      thumbs:
        items:
          $ref: '#/definitions/metadata.Thumb'
        type: array
      title:
        type: string
      top250:
        type: string
      trailer:
        type: string
      tvshow:
        $ref: '#/definitions/metadata.TVShowFields'
      user_rating:
        type: string
      year:
        type: integer
    type: object
  metadata.MovieFields:
    properties:
      countries:
        items:
          type: string
        type: array
      fanart:
        $ref: '#/definitions/metadata.Fanart'
      original_language:
        type: string
      set:
        $ref: '#/definitions/metadata.MovieSet'
      show_links:
        items:
          type: string
        type: array
      sort_title:
        type: string
    type: object
  metadata.MovieSet:
    properties:
      name:
        type: string
      overview:
        type: string
    type: object
  metadata.MusicVideoFields:
    properties:
      album:
        type: string
      artists:
        items:
          type: string
        type: array
      track:
        type: string
    type: object
  metadata.Rating:
    properties:
      default:
        type: boolean
      max:
        type: string
      name:
        type: string
      value:
        type: string
      votes:
        type: string
    type: object
  metadata.Ratings:
    properties:
      rating:
        items:
          $ref: '#/definitions/metadata.Rating'
        type: array
    type: object
  metadata.Resume:
    properties:
      position:
        type: string
      total:
        type: string
    type: object
  metadata.SeasonSummary:
    properties:
      folder_name:
        type: string
      named_season:
        type: string
      season_number:
        type: integer
      season_plot:
        type: string
    type: object
  metadata.TVShowFields:
    properties:
      seasons:
        items:
          $ref: '#/definitions/metadata.SeasonSummary'
        type: array
    type: object
  metadata.Thumb:
    properties:
      aspect:
        type: string
      preview:
        type: string
      season:
        type: string
      type:
        type: string
      value:
        type: string
    type: object
  metadata.UnknownElement:
    properties:
      attrs:
        items:
          $ref: '#/definitions/xml.Attr'
        type: array
      inner_xml:
        type: string
      name:
        type: string
    type: object
  metadata.VideoAttributes:
    properties:
      air_date:
        $ref: '#/definitions/types.Date'
      edition:
        type: string
      episode_numbers:
        items:
          type: integer
        type: array
      episode_title:
        type: string
      extra_source:
        type: string
      extra_type:
        description: ExtraType and ExtraSource are set only on extra videos.
        type: string
      is_special:
        type: boolean
      season_number:
        description: |-
          SeasonNumber is nil when it could not be resolved, which is different
          from season zero (specials).
        type: integer
      stack_number:
        type: integer
      stack_type:
        type: string
      three_d_format:
        type: string
      title:
        type: string
      version_label:
        type: string
      year:
        type: integer
    type: object
...
  scanmodel.AudioStream:
    properties:
      channels:
        type: string
      codec:
        type: string
      language:
        type: string
    type: object
  scanmodel.DirectoryType:
    enum:
    - movie
    - tv
    - music_video
    type: string
    x-enum-varnames:
    - TypeMovie
    - TypeTV
    - TypeMusicVideo
  scanmodel.FileInfo:
    properties:
      stream_details:
        $ref: '#/definitions/scanmodel.StreamDetails'
    type: object
  scanmodel.FileStat:
    properties:
      accessed_at:
        type: string
      block_size:
        type: integer
      blocks:
        type: integer
      changed_at:
        type: string
      created_at:
        description: |-
          CreatedAt is the file's birth time. It stays zero on Linux, which does
          not expose one through stat(2).
        type: string
      device_id:
        type: integer
      gid:
        type: integer
      group_name:
        type: string
      inode:
        type: integer
      is_symlink:
        description: |-
          IsSymlink reports that this record describes a symbolic link rather than
          the file it points at. The scan walks with lstat, so a symlinked sidecar
          reports the link's own stat and says so here instead of quietly reporting
          its target's.
        type: boolean
      link_count:
        type: integer
      mode:
        description: |-
          Mode is the symbolic rendering, "-rw-r--r--", and ModeBits the numeric
          permission bits alongside setuid, setgid and sticky — which is why the
          raw stat mode is masked rather than read through fs.FileMode.Perm, since
          that drops the three special bits.
        type: string
      mode_bits:
        type: integer
      modified_at:
        type: string
      owner_name:
        type: string
      size_bytes:
        type: integer
      uid:
        type: integer
    type: object
  scanmodel.ImageInfo:
    properties:
      codec:
        type: string
      error:
        description: |-
          Error records why the header could not be read, for a file named like
          artwork that cannot be decoded. It is the only field set when it is set,
          so a corrupt poster stays distinguishable from one never examined.
        type: string
      height:
        type: integer
      width:
        type: integer
    type: object
  scanmodel.MediaFile:
    properties:
      directory_id:
        type: string
      directory_path:
        type: string
      extension:
        type: string
      file_info:
        $ref: '#/definitions/scanmodel.FileInfo'
      file_name:
        type: string
      id:
        type: string
      metadata:
        $ref: '#/definitions/metadata.Metadata'
      modified_at:
        type: string
      path:
        type: string
      record_type:
        $ref: '#/definitions/scanmodel.RecordType'
      relative_path:
        type: string
      scan_root_path:
        type: string
      scanned_at:
        type: string
      sidecars:
        items:
          $ref: '#/definitions/scanmodel.SidecarFile'
        type: array
      size_bytes:
        type: integer
      stat:
        allOf:
        - $ref: '#/definitions/scanmodel.FileStat'
        description: |-
          Stat is the filesystem's view of the file — ownership, permissions, inode,
          timestamps — as recorded at scan time. FileInfo below is the codec's view;
          the two describe different things despite the similar names.
      type:
        $ref: '#/definitions/scanmodel.DirectoryType'
      video:
        $ref: '#/definitions/metadata.VideoAttributes'
      warnings:
        items:
          type: string
        type: array
    type: object
  scanmodel.RecordType:
    enum:
    - tvseries
    - media_file
    type: string
    x-enum-varnames:
    - RecordTypeTVSeries
    - RecordTypeMediaFile
  scanmodel.SidecarCategory:
    enum:
    - image
    - video_extra
    - subtitle
    - metadata
    - audio
    - disc_structure
    - trickplay
    - unknown
    type: string
    x-enum-varnames:
    - SidecarCategoryImage
    - SidecarCategoryVideoExtra
    - SidecarCategorySubtitle
    - SidecarCategoryMetadata
    - SidecarCategoryAudio
    - SidecarCategoryDiscStructure
    - SidecarCategoryTrickplay
    - SidecarCategoryUnknown
  scanmodel.SidecarFile:
    properties:
      category:
        $ref: '#/definitions/scanmodel.SidecarCategory'
      extension:
        type: string
      file_name:
        type: string
      image:
        allOf:
        - $ref: '#/definitions/scanmodel.ImageInfo'
        description: |-
          Image is the codec and dimensions read from the file's header, present
          only on sidecars whose extension marks them as artwork.
      modified_at:
        type: string
      relative_path:
        type: string
      sidecar_type:
        $ref: '#/definitions/scanmodel.SidecarType'
      size_bytes:
        type: integer
      stat:
        allOf:
        - $ref: '#/definitions/scanmodel.FileStat'
        description: |-
          Stat is what the filesystem reports about the file itself: ownership,
          permissions, inode and the full set of timestamps.
      trickplay:
        allOf:
        - $ref: '#/definitions/scanmodel.TrickplayInfo'
        description: |-
          Trickplay places a scrubbing preview tile within its media file's
          previews, and is present only on those.
    type: object
  scanmodel.SidecarType:
    enum:
    - poster
    - fanart
    - banner
    - clearlogo
    - clearart
    - discart
    - thumb
    - landscape
    - trailer
    - behind_the_scenes
    - deleted_scene
    - featurette
    - interview
    - scene
    - short
    - other_extra
    - subtitle
    - nfo
    - theme
    - disc_structure
    - trickplay
    - unknown
    type: string
    x-enum-varnames:
    - SidecarPoster
    - SidecarFanart
    - SidecarBanner
    - SidecarClearLogo
    - SidecarClearArt
    - SidecarDiscArt
    - SidecarThumb
    - SidecarLandscape
    - SidecarTrailer
    - SidecarBehindTheScenes
    - SidecarDeletedScene
    - SidecarFeaturette
    - SidecarInterview
    - SidecarScene
    - SidecarShort
    - SidecarOtherExtra
    - SidecarSubtitle
    - SidecarNFO
    - SidecarTheme
    - SidecarDiscStructure
    - SidecarTrickplay
    - SidecarUnknown
  scanmodel.StreamDetails:
    properties:
      audio:
        items:
          $ref: '#/definitions/scanmodel.AudioStream'
        type: array
      subtitle:
        items:
          $ref: '#/definitions/scanmodel.SubtitleStream'
        type: array
      video:
        items:
          $ref: '#/definitions/scanmodel.VideoStream'
        type: array
    type: object
  scanmodel.SubtitleStream:
    properties:
      language:
        type: string
    type: object
  scanmodel.TVSeason:
    properties:
      folder_name:
        type: string
      season_number:
        type: integer
      sidecars:
        items:
          $ref: '#/definitions/scanmodel.SidecarFile'
        type: array
    type: object
  scanmodel.TVSeries:
    properties:
      folder_name:
        type: string
      id:
        type: string
      metadata:
        $ref: '#/definitions/metadata.Metadata'
      path:
        type: string
      record_type:
        $ref: '#/definitions/scanmodel.RecordType'
      scan_root_path:
        type: string
      scanned_at:
        type: string
      seasons:
        items:
          $ref: '#/definitions/scanmodel.TVSeason'
        type: array
      sidecars:
        items:
          $ref: '#/definitions/scanmodel.SidecarFile'
        type: array
      type:
        $ref: '#/definitions/scanmodel.DirectoryType'
      warnings:
        items:
          type: string
        type: array
    type: object
  scanmodel.TrickplayInfo:
    properties:
      tile_height:
        type: integer
      tile_index:
        description: |-
          TileIndex is nil for a file inside a resolution folder that is not a
          numbered tile, which is different from being the first one.
        type: integer
      tile_width:
        type: integer
      width:
        type: integer
    type: object
  scanmodel.VideoStream:
    properties:
      aspect:
        type: string
      codec:
        type: string
      duration_in_seconds:
        type: string
      hdr_type:
        type: string
      height:
        type: string
      stereo_mode:
        type: string
      width:
        type: string
    type: object
  types.Date:
    properties:
      time.Time:
        type: string
    type: object
  xml.Attr:
    properties:
      name:
        $ref: '#/definitions/xml.Name'
      value:
        type: string
    type: object
  xml.Name:
    properties:
      local:
        type: string
      space:
        type: string
    type: object
info:
  contact: {}
  description: Event-driven Go API backed by MongoDB and Redis. Every request carries
    a correlation ID (X-Correlation-ID), generated by the server if the caller doesn't
    supply one.
  title: Metarr API
  version: "1.0"
paths:
...
  /api/local-directories:
    get:
      description: Returns the directory records in the local_directory collection,
        optionally filtered by media type and scan root path.
      parameters:
      - description: 'Filter by media type: movie, tv or music_video'
        in: query
        name: type
        type: string
      - description: Filter by the configured scan directory these were found under
        in: query
        name: scan_root
        type: string
      - description: Maximum records to return (default 100, max 500)
        in: query
        name: limit
        type: integer
      - description: Records to skip
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/scanmodel.TVSeries'
            type: array
        "400":
          description: invalid type, limit or skip
          schema:
            type: string
      security:
      - ApiKeyHeaderAuth: []
      - ApiKeyQueryAuth: []
      summary: List scanned directories
      tags:
      - Media
  /api/local-directories/{id}:
    get:
      description: Returns one directory record by its generated id.
      parameters:
      - description: Directory record id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scanmodel.TVSeries'
        "400":
          description: malformed id
          schema:
            type: string
        "404":
          description: no directory with that id
          schema:
            type: string
      security:
      - ApiKeyHeaderAuth: []
      - ApiKeyQueryAuth: []
      summary: Fetch a scanned directory
      tags:
      - Media
  /api/local-directories/{id}/media-files:
    get:
      description: Returns the media file records — the movie, episode or music video
        files themselves — belonging to one scanned directory.
      parameters:
      - description: Directory record id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/scanmodel.MediaFile'
            type: array
        "400":
          description: malformed id
          schema:
            type: string
        "404":
          description: no directory with that id
          schema:
            type: string
      security:
      - ApiKeyHeaderAuth: []
      - ApiKeyQueryAuth: []
      summary: List a directory's media files
      tags:
      - Media
  /api/local-directories/{id}/nfo:
    get:
      description: Reads and parses one .nfo file inside a scanned directory, live
        from disk rather than from the scan snapshot. The path is relative to the
        directory and may not escape it.
      parameters:
      - description: Directory record id
        in: path
        name: id
        required: true
        type: string
      - description: Path to the .nfo file, relative to the directory
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/metadata.Metadata'
        "400":
          description: missing or unsafe path, or not an .nfo file
          schema:
            type: string
        "404":
          description: no directory with that id, or no such file
          schema:
            type: string
        "422":
          description: the file could not be parsed
          schema:
            type: string
      security:
      - ApiKeyHeaderAuth: []
      - ApiKeyQueryAuth: []
      summary: Read an NFO file from disk
      tags:
      - Media
...
  /api/media-files/{id}:
    get:
      description: Returns one media file record by its generated id, including its
        own NFO metadata, subtitles, artwork and episode ids.
      parameters:
      - description: Media file record id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scanmodel.MediaFile'
        "400":
          description: malformed id
          schema:
            type: string
        "404":
          description: no media file with that id
          schema:
            type: string
      security:
      - ApiKeyHeaderAuth: []
      - ApiKeyQueryAuth: []
      summary: Fetch a media file record
      tags:
      - Media
...
securityDefinitions:
  ApiKeyHeaderAuth:
    description: API key via the X-Api-Key header.
    in: header
    name: X-Api-Key
    type: apiKey
  ApiKeyQueryAuth:
    description: API key via the apikey query parameter. Either this or ApiKeyHeaderAuth
      satisfies auth on protected routes.
    in: query
    name: apikey
    type: apiKey
swagger: "2.0"